requirements.txt
# ═══════════════════════════════════════════════════
# Core - LangChain
# ═══════════════════════════════════════════════════
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langchain-core==0.2.40
openai==1.50.2

# ═══════════════════════════════════════════════════
# Web framework
# ═══════════════════════════════════════════════════
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# ═══════════════════════════════════════════════════
# Database
# ═══════════════════════════════════════════════════
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
pgvector==0.3.4

# ═══════════════════════════════════════════════════
# Telemetry
# ═══════════════════════════════════════════════════
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0

# ═══════════════════════════════════════════════════
# Utilities
# ═══════════════════════════════════════════════════
python-dotenv==1.0.1
httpx==0.27.2
structlog==24.4.0
python-dateutil==2.9.0
jinja2==3.1.4

# FIXED: langchain 0.2.16 requires tenacity >=8.1.0,<9.0.0
# Was: tenacity==9.0.0 (INCOMPATIBLE)
# Now: tenacity==8.5.0 (compatible with langchain's constraint)
tenacity==8.5.0

# ═══════════════════════════════════════════════════
# Testing
# ═══════════════════════════════════════════════════
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
aiosqlite==0.20.0